Added a tree builder function that grabs a Json style variable
